==> bin/nb_00.py <==

#################################################
### THIS FILE WAS AUTOGENERATED! DO NOT EDIT! ###
#################################################
# file to edit: dev_nb/00_exports_helper_functions.ipynb
from pathlib import Path
import json,fire,re
import pandas as pd
import numpy as np
import pickle, gzip, math
import matplotlib as mpl
import matplotlib.pyplot as plt
import sys
import csv
import os
from timeit import default_timer as timer
import operator
import argparse
import subprocess
import glob
import time

TEST = 'test'

def test(a,b,cmp,cname=None):
    if cname is None: cname=cmp.__name__
    assert cmp(a,b),f'{cname}:\n{a}\n{b}'

def test_eq(a,b): test(a,b,operator.eq,'==')


def makeFolders(folder_list):
    for directory in folder_list:
        if not directory: # Make sure not an empty string ""
            continue
        else:
            if not os.path.exists(directory):
                os.makedirs(directory)

def removeFile(filename):
    try:
        os.remove(filename)
    except OSError:
        pass

def removeAndZip(f):
    f = Path(f)
    z = f'{f}.gz'
    try:
        os.remove(z)
    except OSError:
        pass
    subprocess.run(f"gzip {f}", shell=True)


def fix_bool(s):
    # Nextflow messes up the false flag
    if (s == "false" or s== "False"):
        s = False
    if (s == "true" or s == "True"):
        s = True
    return s
